sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openos"
version = "0.1.0"
description = "Python SDK for controlling OpenOS desktop VMs"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "Pillow>=10.0",
]

[project.urls]
Repository = "https://github.com/Emericen/OpenOS"
